Add OptionPage component and moved playlistContainer logic into it
